Moved thread-unsafe DatabaseUrl code to its initializer
